requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
gpytorch>=1.11.0
z3-solver>=4.12.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
netzob
pandas>=2.0.0
tensorboard>=2.13.0
pyyaml>=6.0
